api/requirements.txt
markovify==0.9.4
Unidecode==1.4.0
requests==2.33.1
urllib3==2.6.3
flask==3.1.3
python-dotenv==1.2.2
gunicorn==25.3.0
flask-cors==6.0.2
boto3==1.42.86
